cart

+++ app/src/main/java/ipleiria/pt/amsi/vapeshop/model/Product.java
@@ -10,14 +10,14 @@ public class Product implements Saleable {
     private  long id;
     private String name ;
     private String description;
-    private Float price;
+    private BigDecimal price;
     private String image_url;
 
     public Product() {
         super();
     }
 
-    public Product(long id, String name, String description, Float price, String image_url) {
+    public Product(long id, String name, String description, BigDecimal price, String image_url) {
         this.id = id;
         this.name = name;
         this.description = description;
@@ -37,7 +37,7 @@ public class Product implements Saleable {
         return description;
     }
 
-    public Float getPrice() {
+    public BigDecimal getPrice() {
         return price;
     }
 
@@ -57,7 +57,7 @@ public class Product implements Saleable {
         this.description = description;
     }
 
-    public void setPrice(Float price) {
+    public void setPrice(BigDecimal price) {
         this.price = price;
     }
